+++ vkcs/backup/data_source_vkcs_backup_providers_test.go
@@ -1,0 +1,27 @@
+package backup_test
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/acctest"
+)
+
+func TestAccBackupProvidersDataSource_basic(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() { acctest.AccTestPreCheck(t) },
+		Steps: []resource.TestStep{
+			{
+				ProtoV6ProviderFactories: acctest.AccTestProtoV6ProviderFactories,
+				Config:                   testAccBackupProvidersDataSourceBasic,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("data.vkcs_backup_providers.providers", "providers.#", "2"),
+				),
+			},
+		},
+	})
+}
+
+const testAccBackupProvidersDataSourceBasic = `
+data "vkcs_backup_providers" "providers" {}
+`